"Fix" issues 1-3 by including the router
The dependency graph was broken because the state of the JhOrgDashboard
lacked data for its child router (it is in its initial data but that has
no effect because we "load" the data for the component dynamically via
merge-component and the router data was not included).

However this "fix" is not a prod solution b/c here we just add it to the
"incoming" data. In a proper solution we would likely add it in
pre-merge - but I failed to find out how to do that, so far.