Slow deferred parent issue
Here the parent of a report router - JhOrgDashboard - takes 15s to load.
But routing to the child starts too and is finished first - and we get
this warning:

> get-ident was invoked on  com.example.ui.item-forms/InventoryReport  with nil props (this could mean it wasn't yet mounted)

logged in the console (of course it is not mounted yet, b/c the parent
is not rendering its body yet).

Note: I am directly accessing http://localhost:3000/jh-org-dashboard/123/item-inventory-report?_rp_=WyJeICIsIn46ZmFrZS9vcmctbnIiLDEyM10%3D&

Reminder: MainRouter has the target JhOrgDashboard, which includes the
JhOrgDetailsRouter which has the report as a target.

Transactions of interest (earlierst first):

1. State Machine (BEGIN):com.example.ui/MainRouter :target [:fake/org-nr "123"]
2. dr/target-ready {:target [:fake/org-nr "123"]})
3. State Machine (BEGIN):com.example.ui/JhOrgDetailsRouter
4. State Machine (BEGIN)[:com.fulcrologic.rad.report/id :com.example.ui.item-forms/InventoryReport]
5. target-ready {:target [:com.fulcrologic.rad.report/id .../InventoryReport]}) [I assume here we get the warning]
6. MainRouter :ready!
7. LOAD report items
8. JhOrgDetailsRouter :ready!
9. InventoryReport :event/loaded