Issue 1: Broken reload on accounts/accounts
Going to the URL http://me:3000/accounts/accounts?_rp_=WyJeICJd&
is broken - it will display "Welcome to the Demo. .." with a loading
indicator, forever. In Console it logs:

> ERROR dr/target-ready! was called but there was no router waiting for the target listed:  [:component/id :com.example.ui/Accounts] This could mean you sent one ident, and indicated ready on another.

This is different from the problem I expected but still worth exploring.
The cause is 99% surely the newly added `:will-enter` in `Accounts`.